requirements.txt
streamlit
opencv-python
numpy
pillow
gitpython
torch
torchvision
flask
